requirements.txt
# AIfred Intelligence - Reflex Version
# Multi-Backend LLM Chat Assistant with PWA support

# Core Framework
reflex>=0.8.17
pydantic>=2.0.0         # For data models (required by Reflex State)

# LLM Backends
httpx>=0.28.0           # For Ollama HTTP API
fastapi>=0.115.0        # REST API for external integrations
openai>=1.0.0           # For vLLM (OpenAI-compatible API)
pynvml>=11.0.0          # NVIDIA GPU monitoring for VRAM detection
psutil>=5.9.0           # System resource monitoring (RAM for hybrid mode)

# Web Search & Scraping
trafilatura>=1.13.0     # Web content extraction (includes lxml)
playwright>=1.50.0      # JavaScript-heavy page scraping
requests>=2.31.0        # HTTP requests for search APIs (Brave, SearXNG)
pymupdf>=1.26.0         # PDF text extraction (AWMF guidelines, PubMed PDFs)

# Vector Database & Semantic Search
chromadb>=0.4.24        # Vector database for semantic caching
ollama>=0.4.0           # Ollama Python client for embeddings
numpy>=1.24.0           # Required by chromadb embeddings

# Audio Processing (STT/TTS)
edge-tts==7.2.1         # Text-to-Speech (7.2.3 has NoAudioReceived bug)
dashscope>=1.24.6       # DashScope Qwen3-TTS Cloud API
openai-whisper>=20231117 # Speech-to-Text

# Configuration & Environment
python-dotenv>=1.0.0    # .env file support

# Model Management
huggingface_hub>=1.4.1          # HuggingFace model downloads (hf CLI + GGUF discovery)
hf>=1.4.1                       # HuggingFace CLI tool (hf download, hf upload, etc.)
gguf>=0.10.0                    # GGUF metadata reading (native context detection)

# Document Processing
chardet>=5.0.0          # Automatic encoding detection for uploaded documents
python-docx>=1.0.0      # Microsoft Word (.docx) text extraction
openpyxl>=3.1.0         # Microsoft Excel (.xlsx) text extraction
python-pptx>=1.0.0      # Microsoft PowerPoint (.pptx) text extraction
odfpy>=1.4.0            # LibreOffice/OpenDocument (.odt, .ods, .odp) text extraction

# Utilities
python-multipart>=0.0.20
platformdirs>=4.5.0
mistune>=3.0.0          # Markdown-to-HTML for Chat Export

# Development
pytest>=8.0.0
black>=24.0.0
Pillow>=10.0.0
